refactor(engine): extract synthetic-shadow lookup out of the engine-core
