+++ packages/@lwc/engine/src/framework/stylesheet.ts
@@ -8,8 +8,10 @@ import { assert, create, emptyString, forEach, isArray, isUndefined } from '@lwc
 import { VNode } from '../3rdparty/snabbdom/types';
 
 import * as api from './api';
-import { EmptyArray, useSyntheticShadow } from './utils';
+import { EmptyArray } from './utils';
 import { VM } from './vm';
+
+import { useSyntheticShadow } from '../../dom/src/env/shadow-dom';
 import { removeAttribute, setAttribute } from '../../dom/src/env/element';
 /**
  * Function producing style based on a host and a shadow selector. This function is invoked by
